Move worker_config_buffer.hpp out of api/tt-metalium
worker_config_buffer.hpp is an internal dispatch concept, so it doesn't belong in api/tt-metalium.

Move it into tt_metal_impl/dispatch, alongside worker_config_buffer.cpp.
