+++ tt_metal/CMakeLists.txt
@@ -103,7 +103,6 @@ target_sources(
             api/tt-metalium/utils.hpp
             api/tt-metalium/work_executor_types.hpp
             api/tt-metalium/work_split.hpp
-            api/tt-metalium/worker_config_buffer.hpp
             api/tt-metalium/distributed_context.hpp
 )
