+++ tt_metal/api/tt-metalium/command_queue.hpp
@@ -9,7 +9,6 @@
 #include <memory>
 #include <thread>
 
-#include <tt-metalium/worker_config_buffer.hpp>
 #include <tt-metalium/trace_buffer.hpp>
 #include <tt-metalium/command_queue_interface.hpp>
 
@@ -21,6 +20,7 @@ class Event;
 class Program;
 class Kernel;
 class SystemMemoryManager;
+class WorkerConfigBufferMgr;
 
 class CommandQueue {
 public:
