+++ tt_metal/impl/dispatch/worker_config_buffer.cpp
@@ -2,10 +2,11 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+#include "worker_config_buffer.hpp"
+
 #include <assert.hpp>
 #include <stdint.h>
 #include <stdio.h>
-#include <worker_config_buffer.hpp>
 #include <algorithm>
 #include <utility>
 #include <vector>
