+++ tt_metal/impl/trace/dispatch.hpp
@@ -7,13 +7,13 @@
 #include <device.hpp>
 #include <stdint.h>
 #include <trace_buffer.hpp>
-#include <worker_config_buffer.hpp>
 #include <cstddef>
 #include <unordered_map>
 #include <vector>
 
 #include "core_coord.hpp"
 #include "dispatch/dispatch_settings.hpp"
+#include "dispatch/worker_config_buffer.hpp"
 #include "sub_device_types.hpp"
 
 namespace tt {
